ref : recipe recommend api
